refactor: move readback helpers out of TextureProcessor
