+++ Editor/Common/Utils/TextureReadback.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 94aab976597a169db5e8db9bf21c87e8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
