+++ Editor/TextureCompressor/Core/Services/TextureProcessor.cs
@@ -8,9 +8,6 @@ namespace dev.limitex.avatar.compressor.editor.texture
     /// </summary>
     public class TextureProcessor
     {
-        // Lock object for thread-safe RenderTexture operations
-        private static readonly object RenderTextureLock = new object();
-
         private readonly int _minResolution;
         private readonly int _maxResolution;
         private readonly bool _forcePowerOfTwo;
@@ -106,7 +103,7 @@ namespace dev.limitex.avatar.compressor.editor.texture
             bool isNormalMap
         )
         {
-            lock (RenderTextureLock)
+            lock (TextureReadback.RenderTextureLock)
             {
                 try
                 {
@@ -126,15 +123,6 @@ namespace dev.limitex.avatar.compressor.editor.texture
                     return null;
                 }
             }
-        }
-
-        internal static void CopyTextureSettings(Texture2D source, Texture2D dest)
-        {
-            dest.wrapModeU = source.wrapModeU;
-            dest.wrapModeV = source.wrapModeV;
-            dest.wrapModeW = source.wrapModeW;
-            dest.filterMode = source.filterMode;
-            dest.anisoLevel = source.anisoLevel;
         }
 
         /// <summary>
@@ -166,7 +154,11 @@ namespace dev.limitex.avatar.compressor.editor.texture
                 }
             }
 
-            var readable = BlitToReadable(texture, RenderTextureReadWrite.Linear, linearFlag: true);
+            var readable = TextureReadback.BlitToReadable(
+                texture,
+                RenderTextureReadWrite.Linear,
+                linearFlag: true
+            );
             if (readable == null)
                 return new Color[0];
             try
@@ -176,81 +168,6 @@ namespace dev.limitex.avatar.compressor.editor.texture
             finally
             {
                 Object.DestroyImmediate(readable);
-            }
-        }
-
-        /// <summary>
-        /// Blits a texture into a temporary RenderTexture with the given color-space
-        /// policy and returns a readable RGBA32 copy (no mip chain, settings not
-        /// copied), or null on failure. Owns the RenderTexture lock, the active-RT
-        /// save/restore, and the GL.sRGBWrite guard. Caller destroys the result.
-        /// </summary>
-        internal static Texture2D BlitToReadable(
-            Texture2D source,
-            RenderTextureReadWrite colorSpace,
-            bool linearFlag
-        )
-        {
-            lock (RenderTextureLock)
-            {
-                RenderTexture previous = RenderTexture.active;
-                var previousSRGBWrite = GL.sRGBWrite;
-                RenderTexture rt = null;
-                Texture2D readable = null;
-                try
-                {
-                    // Use explicit RenderTexture lifecycle instead of GetTemporary/ReleaseTemporary
-                    // so that native GPU memory is freed immediately by DestroyImmediate,
-                    // rather than being held in Unity's RT pool across calls.
-                    rt = new RenderTexture(
-                        source.width,
-                        source.height,
-                        0,
-                        RenderTextureFormat.ARGB32,
-                        colorSpace
-                    );
-                    rt.Create();
-
-                    // The write-side linear->sRGB encode into an sRGB RT is gated by
-                    // GL.sRGBWrite (editor IMGUI leaves it false), not by the RT's
-                    // readWrite flag. Harmless for linear RTs (no conversion).
-                    GL.sRGBWrite = true;
-                    Graphics.Blit(source, rt);
-                    RenderTexture.active = rt;
-
-                    readable = new Texture2D(
-                        source.width,
-                        source.height,
-                        TextureFormat.RGBA32,
-                        false,
-                        linearFlag
-                    );
-                    readable.ReadPixels(new Rect(0, 0, source.width, source.height), 0, 0);
-                    readable.Apply(false);
-
-                    var result = readable;
-                    readable = null;
-                    return result;
-                }
-                catch (System.Exception e)
-                {
-                    Debug.LogWarning(
-                        $"[TextureCompressor] Failed to read back texture '{source.name}': {e.Message}"
-                    );
-                    return null;
-                }
-                finally
-                {
-                    GL.sRGBWrite = previousSRGBWrite;
-                    RenderTexture.active = previous;
-                    if (readable != null)
-                        Object.DestroyImmediate(readable);
-                    if (rt != null)
-                    {
-                        rt.Release();
-                        Object.DestroyImmediate(rt);
-                    }
-                }
             }
         }
     }
